requirements.txt
feedparser==6.0.10
newspaper3k==0.2.8
lxml_html_clean
requests==2.31.0
tenacity==8.2.3
pydantic==2.4.2
pendulum>=2.1.2
python-dotenv==1.0.0
pyyaml==6.0.1
pytest==7.4.3

